==> data_analysis_python/00_scrape_images.ipynb <==
import requests
import sys
%load_ext autotime

r = requests.get('http://jmcauley.ucsd.edu/imageGraph/models/urls_Clothing_Shoes_and_Jewelry.txt').text

len(r)

l = r.split('\n') 
bad = []

for i, pic in enumerate(l):
    if i < 940000:
        pass
    else:
        if i % 10000 == 0:
            sys.stderr.write(str(i))
        try:
            pid, url = pic.split(' ')
            f = open('./pics/{0}.jpg'.format(pid),'wb')
            f.write(requests.get(url).content)
        except:
            bad.append(pic)

import os

len(os.listdir('./pics'))

def getFolderSize(folder):
    total_size = os.path.getsize(folder)
    for item in os.listdir(folder):
        itempath = os.path.join(folder, item)
        if os.path.isfile(itempath):
            total_size += os.path.getsize(itempath)
        elif os.path.isdir(itempath):
            total_size += getFolderSize(itempath)
    return total_size

52073134604./1e9
