==> data_analysis_python/03_image_features.ipynb <==
from pandas import *

import struct

def readImageFeatures(path):
    f = open(path, 'rb')
    while True:
        asin = f.read(10)
        if asin == '':
            break
        feature = []
        for i in range(4096):
            feature.append(struct.unpack('f', f.read(4)))
        yield asin, feature

df = read_csv('/mnt/amazon/data/boots_aws.csv', index_col = 'Unnamed: 0')
asin = df.asin.values

path = '/mnt/amazon/data/image_features_Clothing_Shoes_and_Jewelry.b'
r = readImageFeatures(path)

with open('/mnt/amazon/data/features_text.txt','w') as w:
    for img in r:
        if img[0] in asin:
            w.write(str(img))
            w.write('\n')
