==> data_analysis_python/04_image_features_prep.ipynb <==
from pandas import *
import numpy as np

f = open('/mnt/amazon/data/features_text.txt').readlines()

len(f)

df_dict = {}
for r in f:
    row = r.split("'")
    features = [np.float64(f.replace('(','').replace(',)','')) for f in row[2][3:-3].split(', ')]
    df_dict[row[1]] = features

df = DataFrame.from_dict(df_dict, orient = 'index')
df.index.name = 'asin'

df.shape

df.head()

df.to_csv('/mnt/amazon/data/boots_features.csv')
