==> data_analysis_python/12_gmm_cluster.ipynb <==
import os
from pandas import *
from sklearn.mixture import VBGMM

df = read_csv('../boots_dataset/features_women.csv', index_col = 'asin')

df.head()

model = VBGMM(n_components=25, covariance_type='diag', alpha=1.0, random_state=None, 
      thresh=None, tol=0.001, verbose=0, min_covar=None, n_iter=10, 
      params='wmc', init_params='wmc')

nonzero = np.nonzero([0, 1, 1, 0])[0]

subset_cluster = df.iloc[:, nonzero]

model.fit(subset_cluster)

DataFrame(model.predict(subset_cluster)).loc[:, 0].value_counts()

k = 25

cluster_dict = {}
pred = model.predict(subset_cluster)
df.loc[:, 'cluster'] = pred

for i, cluster in enumerate(np.unique(pred)):
    cluster_dict[i] = list(df[df.cluster == cluster].drop('cluster', axis = 1).index)

from pandas import *
from sklearn.mixture import VBGMM
from sklearn.mixture import DPGMM
from sklearn.cluster import KMeans
import sys

class ClusteringImages:
    def __init__(self, df, coef, model_name, alpha, k):
        self.generate_model(model_name, alpha, k)
        self.features_df = df.iloc[:, np.nonzero(coef)[0]]
        self.fit_model()
        self.cluster()
        
    def generate_model(self, model_name, alpha, k):
        if model_name == 'VBGMM':
            self.model = VBGMM(n_components=k, covariance_type='diag', alpha=alpha, random_state=None, 
              thresh=None, tol=0.001, verbose=0, min_covar=None, n_iter=10, 
              params='wmc', init_params='wmc')
        elif model_name == 'DPGMM':
            self.model = DPGMM(n_components=k, covariance_type='diag', alpha=alpha, random_state=None, 
                  thresh=None, tol=0.001, verbose=0, min_covar=None, n_iter=10, 
                  params='wmc', init_params='wmc')        
        elif model_name == 'K-Means':
            self.model = KMeans(n_clusters=k, init='k-means++', n_init=10, max_iter=300, 
                                tol=0.0001, precompute_distances='auto', verbose=0, random_state=None, 
                                copy_x=True, n_jobs=1)
        else:
            print 'enter valid cluster'
            sys.exit(0)
        
    def fit_model(self):
        self.model.fit(self.features_df)
    
    def cluster(self):
        self.cluster_dict = {}
        pred = self.model.predict(self.features_df)
        self.features_df.loc[:, 'cluster'] = pred

        for i, cluster in enumerate(np.unique(pred)):
            self.cluster_dict[i] = list(self.features_df[self.features_df.loc[:, 'cluster'] == cluster]\
                                   .drop('cluster', axis = 1)\
                                   .index)

c = ClusteringImages(df, [0, 0, 1, 2, 3, 0, 0, 1], 'VBGMM', 1, 25)
